Made all the quick changes noted in the issue from the 4/1 feedback meeting also working keypad
